--- src/main/java/org/runejs/client/sound/Class22.java
@@ -1,15 +1,0 @@
-package org.runejs.client.sound;
-
-public abstract class Class22 {
-    public abstract void method300(byte[] is, boolean bool, int i_17_);
-
-    public abstract void method301(int i, int i_18_);
-
-    public abstract void method302(int i);
-
-    public abstract void method303(byte i);
-
-    public abstract void method304(byte i, int i_19_, int i_20_);
-
-    public abstract void method306(byte i);
-}
